+++ src/StudentTracker.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace StudentTracker.Core;
-
-public class Class1
-{
-
-}