+++ src/StudentTracker.Data/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace StudentTracker.Data;
-
-public class Class1
-{
-
-}